% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{GroupList}
\alias{GroupList}
\title{GroupList Object}
\usage{
GroupList(id = NULL, items = NULL, nextPageToken = NULL,
  selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{[Output Only] A list of Group resources}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server defined URL for this resource}
}
\value{
GroupList object
}
\description{
GroupList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

