% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{GroupsAddMemberRequest}
\alias{GroupsAddMemberRequest}
\title{GroupsAddMemberRequest Object}
\usage{
GroupsAddMemberRequest(users = NULL)
}
\arguments{
\item{users}{Fully-qualified URLs of the User resources to add}
}
\value{
GroupsAddMemberRequest object
}
\description{
GroupsAddMemberRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GroupsAddMemberRequest functions: \code{\link{groups.addMember}}
}

