% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{GroupsRemoveMemberRequest}
\alias{GroupsRemoveMemberRequest}
\title{GroupsRemoveMemberRequest Object}
\usage{
GroupsRemoveMemberRequest(users = NULL)
}
\arguments{
\item{users}{Fully-qualified URLs of the User resources to remove}
}
\value{
GroupsRemoveMemberRequest object
}
\description{
GroupsRemoveMemberRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other GroupsRemoveMemberRequest functions: \code{\link{groups.removeMember}}
}
