% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_functions.R
\name{linux.getLinuxAccountViews}
\alias{linux.getLinuxAccountViews}
\title{Retrieves a list of user accounts for an instance within a specific project.}
\usage{
linux.getLinuxAccountViews(project, zone, instance, filter = NULL,
  maxResults = NULL, orderBy = NULL, pageToken = NULL)
}
\arguments{
\item{project}{Project ID for this request}

\item{zone}{Name of the zone for this request}

\item{instance}{The fully-qualified URL of the virtual machine requesting the views}

\item{filter}{Sets a filter expression for filtering listed resources, in the form filter={expression}}

\item{maxResults}{The maximum number of results per page that should be returned}

\item{orderBy}{Sorts list results by a certain order}

\item{pageToken}{Specifies a page token to use}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud-platform.read-only
\item https://www.googleapis.com/auth/cloud.useraccounts
\item https://www.googleapis.com/auth/cloud.useraccounts.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud-platform.read-only, https://www.googleapis.com/auth/cloud.useraccounts, https://www.googleapis.com/auth/cloud.useraccounts.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/compute/docs/access/user-accounts/api/latest/}{Google Documentation}
}

