% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LinuxAccountViews}
\alias{LinuxAccountViews}
\title{LinuxAccountViews Object}
\usage{
LinuxAccountViews(groupViews = NULL, userViews = NULL)
}
\arguments{
\item{groupViews}{[Output Only] A list of all groups within a project}

\item{userViews}{[Output Only] A list of all users within a project}
}
\value{
LinuxAccountViews object
}
\description{
LinuxAccountViews Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of all Linux accounts for this project. This API is only used by Compute Engine virtual machines to get information about user accounts for a project or instance. Linux resources are read-only views into users and groups managed by the Compute Engine Accounts API.
}

