% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{LinuxGetLinuxAccountViewsResponse}
\alias{LinuxGetLinuxAccountViewsResponse}
\title{LinuxGetLinuxAccountViewsResponse Object}
\usage{
LinuxGetLinuxAccountViewsResponse(resource = NULL)
}
\arguments{
\item{resource}{[Output Only] A list of authorized user accounts and groups}
}
\value{
LinuxGetLinuxAccountViewsResponse object
}
\description{
LinuxGetLinuxAccountViewsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

