% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{User}
\alias{User}
\title{User Object}
\usage{
User(creationTimestamp = NULL, description = NULL, groups = NULL,
  id = NULL, name = NULL, owner = NULL, publicKeys = NULL,
  selfLink = NULL)
}
\arguments{
\item{creationTimestamp}{[Output Only] Creation timestamp in RFC3339 text format}

\item{description}{An optional textual description of the resource; provided by the client when the resource is created}

\item{groups}{[Output Only] A list of URLs to Group resources who contain the user}

\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{name}{Name of the resource; provided by the client when the resource is created}

\item{owner}{Email address of account's owner}

\item{publicKeys}{[Output Only] Public keys that this user may use to login}

\item{selfLink}{[Output Only] Server defined URL for the resource}
}
\value{
User object
}
\description{
User Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A User resource.
}
\seealso{
Other User functions: \code{\link{users.insert}}
}
