% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouduseraccounts_objects.R
\name{UserList}
\alias{UserList}
\title{UserList Object}
\usage{
UserList(id = NULL, items = NULL, nextPageToken = NULL, selfLink = NULL)
}
\arguments{
\item{id}{[Output Only] Unique identifier for the resource; defined by the server}

\item{items}{[Output Only] A list of User resources}

\item{nextPageToken}{[Output Only] A token used to continue a truncated list request}

\item{selfLink}{[Output Only] Server defined URL for this resource}
}
\value{
UserList object
}
\description{
UserList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

